Fix narrowing conversion warnings
- bistream.c: use uint32_t for offset (matches size type)
- connection_dtls.c: cast ssize_t to int for BIO_write
- connection_tcp.c: cast ssize_t to int for recv_size subtraction
- connection_udp.c: cast ssize_t/gsize to int for processor calls,
  use ssize_t for sendtofrom return value
- dionaea.c: use unsigned int for ev_backend, cast rlim_cur to int,
  use long for sysconf return value
- incident.c: use guint for g_hash_table_size return
- log.c: cast ev_now (double) to time_t
- processor.c: cast ev_now (double) to time_t
- ssl.c: use long for time() return value

+++ src/connection_dtls.c
@@ -420,7 +420,7 @@ void connection_dtls_io_in_cb(struct ev_loop *loop, struct ev_io *w, int revents
 		{
 			g_critical("Invalid connection for DTLS");
 		}
-		BIO_write(peer->transport.dtls.reading, buf, ret);
+		BIO_write(peer->transport.dtls.reading, buf, (int)ret);
 
 		switch( peer->state )
 		{
